pyproject.toml
[tool.poetry]
name = "nepali-wallets"
version = "0.1.0"
description = "An SDK for payment gateways integration in Nepal."
authors = ["Sudip Ghimire <ghimiresdp@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "nepali_wallets" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
python-decouple = "^3.7"
black = "^23.12.1"


[tool.poetry.group.peer.dependencies]
cryptography = "^41.0.7"

[tool.black]
line-length = 120
target-version = ['py311']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
